Move more widgets to ui lib
