+++ src/ui/widgets/aboutwidget.h
@@ -1,0 +1,44 @@
+/*
+#
+# Friction - https://friction.graphics
+#
+# Copyright (c) Friction contributors
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+# See 'README.md' for more information.
+#
+*/
+
+#ifndef ABOUTWIDGET_H
+#define ABOUTWIDGET_H
+
+#include "ui_global.h"
+
+#include <QWidget>
+
+class UI_EXPORT AboutWidget : public QWidget
+{
+    Q_OBJECT
+public:
+    struct AboutWidgetTab {
+        QString title;
+        QString path;
+        bool html = true;
+    };
+    explicit AboutWidget(QWidget *parent = nullptr);
+
+};
+
+#endif // ABOUTWIDGET_H
