+++ src/ui/widgets/widgetstack.cpp
